Uploaded file for ControlPanel
This is the Control Panel blueprint, it is self-contained and does not require any other files to be placed. The blueprint's physical assets include a back panel, an interface and a trigger volume placed in front of the panel.

Actual interaction with this object requires the relevant scripting within the level blueprint.